Fix : UI Pet when tapped

+++ macnivision/Model/Pet.swift
@@ -11,7 +11,7 @@ struct Pet: Hashable, Codable, Identifiable{
     var id : Int
     var recordId : String
     var userId : String
-    var hungryLevel : Int
+    var hungryLevel : Float
     var foodAmount : Int
     var typePet: TypePet
     enum TypePet: String, CaseIterable, Codable{
